backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
yt-dlp==2025.6.9
pillow==11.3.0
requests==2.31.0
ffmpeg-python==0.2.0